+++ cmd/shortener/main.go
@@ -2,8 +2,13 @@ package main
 
 import "github.com/DyadyaRodya/go-shortener/internal/app"
 
+const (
+	defaultBaseShortURL  = "http://localhost:8080/"
+	defaultServerAddress = `:8080`
+)
+
 func main() {
-	server := app.NewApp("http://localhost:8080/", `:8080`)
+	server := app.NewApp(defaultBaseShortURL, defaultServerAddress)
 
 	err := server.Run()
 	if err != nil {
