+++ env.dist
@@ -1,0 +1,2 @@
+SERVER_ADDRESS=localhost:8080
+BASE_URL=http://localhost:8080/
